refactor: split ts expression parser
