+++ crates/tsv_ts/src/parser/expression.rs
@@ -1,24 +1,20 @@
 // Expression parsing using Pratt parser for operator precedence
 
 use crate::ast::internal::{
-    ArrayPattern, ArrowFunctionBody, ArrowFunctionExpression, AssignmentExpression,
-    AssignmentPattern, AwaitExpression, BinaryExpression, BinaryOperator, BlockStatement,
+    AssignmentExpression, AwaitExpression, BinaryExpression, BinaryOperator, BlockStatement,
     CallExpression, ConditionalExpression, Expression, Identifier, ImportExpression, JsdocCast,
-    Literal, LiteralValue, MemberExpression, MetaProperty, NewExpression, ObjectPattern,
-    ObjectPatternProperty, ObjectProperty, Property, RegexLiteral, RestElement, SequenceExpression,
-    SpreadElement, Super, TSAsExpression, TSInstantiationExpression, TSNonNullExpression,
-    TSSatisfiesExpression, TSTypeAssertion, TaggedTemplateExpression, TemplateElement,
-    TemplateLiteral, ThisExpression, UnaryExpression, UnaryOperator, UpdateExpression,
-    UpdateOperator, YieldExpression,
+    Literal, LiteralValue, MemberExpression, MetaProperty, NewExpression, RegexLiteral,
+    SequenceExpression, SpreadElement, Super, TSAsExpression, TSInstantiationExpression,
+    TSNonNullExpression, TSSatisfiesExpression, TSTypeAssertion, TaggedTemplateExpression,
+    ThisExpression, UnaryExpression, UnaryOperator, UpdateExpression, UpdateOperator,
+    YieldExpression,
 };
 use crate::lexer::{KeywordKind, TokenKind};
 use tsv_lang::{ParseError, Span};
 
 use super::Parser;
-use super::expression_lookahead::{
-    scan_angle_brackets, scan_identifier_then_arrow, scan_parens_then_arrow,
-};
-use super::scan::{parse_number_literal, skip_whitespace_and_comments};
+use super::expression_lookahead::scan_parens_then_arrow;
+use super::scan::parse_number_literal;
 
 //
 // Binding Power Constants for Pratt Parser
@@ -37,32 +33,6 @@ const BP_TS_TYPE_ASSERTION: u8 = 2;
 const BP_YIELD: u8 = 1;
 /// Unary operators (!, -, +, ~, typeof, void, delete, await, ++, --, new)
 const BP_UNARY: u8 = 29;
-
-/// Extract content from template head: `content${ → "content"
-#[inline]
-fn extract_template_head_content(raw: &str) -> &str {
-    if raw.len() >= 3 {
-        &raw[1..raw.len() - 2]
-    } else {
-        ""
-    }
-}
-
-/// Extract content from template tail: }content` → "content"
-#[inline]
-fn extract_template_tail_content(raw: &str) -> &str {
-    if raw.len() >= 2 {
-        &raw[1..raw.len() - 1]
-    } else {
-        ""
-    }
-}
-
-/// Extract content from no-substitution template: `content` → "content"
-#[inline]
-fn extract_template_simple_content(raw: &str) -> &str {
-    extract_template_tail_content(raw) // Same logic: strip first and last char
-}
 
 /// Parsed expression with actual end position tracking
 ///
@@ -1198,77 +1168,6 @@ impl<'a> Parser<'a> {
         is_jsdoc_type_cast_comment(value)
     }
 
-    /// Check if current position starts an arrow function
-    ///
-    /// Scans ahead looking for pattern: `(` ... `)` `=>`
-    fn is_arrow_function_start(&self) -> bool {
-        scan_parens_then_arrow(self.source.as_bytes(), self.current_start)
-    }
-
-    /// Check if current position starts a single-param arrow function: `x =>`
-    ///
-    /// Scans ahead looking for pattern: `identifier` `=>`
-    fn is_single_param_arrow_start(&self) -> bool {
-        scan_identifier_then_arrow(self.source.as_bytes(), self.current_start)
-    }
-
-    /// Check if current position starts a generic arrow function: `<T>() =>`
-    ///
-    /// Scans ahead looking for pattern: `<` ... `>` `(` ... `)` `=>`
-    fn is_generic_arrow_function_start(&self) -> bool {
-        let bytes = self.source.as_bytes();
-        let start = self.current_start;
-
-        // Must start with '<'
-        if start >= bytes.len() || bytes[start] != b'<' {
-            return false;
-        }
-
-        // Scan through type parameters: <T, U extends V, ...>
-        let pos = scan_angle_brackets(bytes, start);
-        if pos == 0 {
-            return false;
-        }
-
-        // After '>', check for `(...) =>` (allow comments: `<T> /* comment */ () =>`)
-        let pos = skip_whitespace_and_comments(bytes, pos);
-        scan_parens_then_arrow(bytes, pos)
-    }
-
-    /// Parse generic arrow function: `<T>() => ...`, `<T, U extends V>() => ...`
-    fn parse_generic_arrow_function(&mut self) -> Result<Expression, ParseError> {
-        let (start, _) = self.current_pos();
-
-        // Parse type parameters: <T, U extends V, ...>
-        let type_parameters = self.parse_type_parameters()?;
-
-        // Capture paren position before parsing params
-        let (params_start, _) = self.current_pos();
-
-        // Parse parameter list
-        let params = self.parse_parameter_list()?;
-
-        // Check for return type annotation: <T>(): type => ... or type predicate
-        let return_type = self.parse_optional_return_type()?;
-
-        self.expect(&TokenKind::Arrow)?; // consume '=>'
-
-        let body = self.parse_arrow_body()?;
-        let end = self.prev_token_end() as u32;
-
-        Ok(Expression::ArrowFunctionExpression(
-            ArrowFunctionExpression {
-                type_parameters: Some(type_parameters),
-                params,
-                body,
-                return_type,
-                r#async: false,
-                params_start: Some(params_start as u32),
-                span: Span::new(start as u32, end),
-            },
-        ))
-    }
-
     /// Parse a TypeScript angle-bracket type assertion: `<Type>expr`
     ///
     /// This is the old-style type assertion syntax. It's equivalent to `expr as Type`
@@ -1761,146 +1660,6 @@ impl<'a> Parser<'a> {
         }))
     }
 
-    /// Parse arrow function body: expression or block statement
-    fn parse_arrow_body(&mut self) -> Result<ArrowFunctionBody, ParseError> {
-        if self.check(&TokenKind::BraceOpen) {
-            let block = self.parse_function_body()?;
-            Ok(ArrowFunctionBody::BlockStatement(block))
-        } else {
-            // Use assignment_expression so comma doesn't consume next object property
-            let expr = self.parse_assignment_expression()?;
-            Ok(ArrowFunctionBody::Expression(Box::new(expr)))
-        }
-    }
-
-    /// Parse arrow function with parentheses: `() => expr` or `(x, y) => expr` or `() => { ... }`
-    ///
-    /// Supports:
-    /// - No parameters: `() => expr`
-    /// - Single parameter: `(x) => expr`
-    /// - Multiple parameters: `(x, y) => expr`
-    /// - Destructuring parameters: `([a, b]) => ...`, `({x, y}) => ...`
-    /// - Default values: `(a = 1) => ...`
-    /// - Expression body: `() => expr`
-    /// - Block body: `() => { ... }`
-    ///
-    /// Note: Single parameter without parens (`x => expr`) is handled by
-    /// `parse_single_param_arrow_function()`.
-    fn parse_arrow_function(&mut self) -> Result<Expression, ParseError> {
-        let (start, _) = self.current_pos();
-
-        // Capture paren position before parsing params
-        let (params_start, _) = self.current_pos();
-
-        // Parse parameter list (reuse shared method)
-        let params = self.parse_parameter_list()?;
-
-        // Check for return type annotation: (): type => ... or type predicate
-        let return_type = self.parse_optional_return_type()?;
-
-        self.expect(&TokenKind::Arrow)?; // consume '=>'
-
-        let body = self.parse_arrow_body()?;
-        let end = self.prev_token_end() as u32;
-
-        Ok(Expression::ArrowFunctionExpression(
-            ArrowFunctionExpression {
-                type_parameters: None, // Generic arrows like <T>() => {} are handled by parse_generic_arrow_function()
-                params,
-                body,
-                return_type,
-                r#async: false, // Non-async arrow function; async ones are parsed via parse_async_arrow_function
-                params_start: Some(params_start as u32),
-                span: Span::new(start as u32, end),
-            },
-        ))
-    }
-
-    /// Parse single-parameter arrow function without parentheses: `x => expr`
-    fn parse_single_param_arrow_function(&mut self) -> Result<Expression, ParseError> {
-        let (start, _) = self.current_pos();
-
-        // Parse the single identifier parameter
-        debug_assert!(matches!(self.current_kind(), TokenKind::Identifier));
-        let (id_start, id_end) = self.current_pos();
-        let symbol = self.intern_identifier();
-        self.advance()?;
-
-        let params = vec![Expression::Identifier(Identifier::simple(
-            symbol,
-            Span::new(id_start as u32, id_end as u32),
-        ))];
-
-        self.expect(&TokenKind::Arrow)?; // consume '=>'
-
-        let body = self.parse_arrow_body()?;
-        let end = self.prev_token_end() as u32;
-
-        Ok(Expression::ArrowFunctionExpression(
-            ArrowFunctionExpression {
-                type_parameters: None,
-                params,
-                body,
-                return_type: None, // Single-param without parens can't have return type
-                r#async: false,
-                params_start: None, // No parens for single-param arrows
-                span: Span::new(start as u32, end),
-            },
-        ))
-    }
-
-    /// Parse async arrow function after 'async' has been consumed: `() => ...`, `x => ...`, or `<T>() => ...`
-    fn parse_async_arrow_function_after_async(
-        &mut self,
-        start: usize,
-    ) -> Result<Expression, ParseError> {
-        // Check for type parameters: `async <T>() => ...`
-        let type_parameters = self.parse_optional_type_parameters()?;
-
-        // Parse parameter list or single parameter
-        // Note: with type parameters, must have parentheses
-        let (params, params_start) = if self.check(&TokenKind::ParenOpen) {
-            let (paren_pos, _) = self.current_pos();
-            (self.parse_parameter_list()?, Some(paren_pos as u32))
-        } else if type_parameters.is_none() && matches!(self.current_kind(), TokenKind::Identifier)
-        {
-            // Single parameter without parens: `async x => ...`
-            // (Not allowed with type parameters)
-            let (id_start, id_end) = self.current_pos();
-            let symbol = self.intern_identifier();
-            self.advance()?;
-            (
-                vec![Expression::Identifier(Identifier::simple(
-                    symbol,
-                    Span::new(id_start as u32, id_end as u32),
-                ))],
-                None,
-            )
-        } else {
-            return Err(self.error_expected_after("'(' or identifier", "async"));
-        };
-
-        // Check for return type annotation or type predicate
-        let return_type = self.parse_optional_return_type()?;
-
-        self.expect(&TokenKind::Arrow)?; // consume '=>'
-
-        let body = self.parse_arrow_body()?;
-        let end = self.prev_token_end() as u32;
-
-        Ok(Expression::ArrowFunctionExpression(
-            ArrowFunctionExpression {
-                type_parameters,
-                params,
-                body,
-                return_type,
-                r#async: true,
-                params_start,
-                span: Span::new(start as u32, end),
-            },
-        ))
-    }
-
     /// Parse a block statement: `{ stmt1; stmt2; }`
     ///
     /// Parses the statements inside a block body (used for function bodies).
@@ -1953,265 +1712,5 @@ impl<'a> Parser<'a> {
             argument: Box::new(argument),
             span: Span::new(start as u32, end),
         }))
-    }
-
-    /// Parse template literal: `hello ${name}`
-    ///
-    /// Handles both simple templates (no interpolation) and templates with expressions.
-    /// See also `parse_template_literal_type()` in statement.rs for type context version.
-    pub(super) fn parse_template_literal(&mut self) -> Result<Expression, ParseError> {
-        let (start, _) = self.current_pos();
-        let mut quasis = Vec::new();
-        let mut expressions = Vec::new();
-
-        match self.current_kind() {
-            TokenKind::NoSubstitutionTemplate => {
-                // Simple template with no interpolation: `hello world`
-                let (elem_start, elem_end) = self.current_pos();
-                let raw = self.current_value();
-                let content = extract_template_simple_content(raw).to_string();
-                let cooked = self.current_decoded().map_or_else(
-                    || Some(content.clone()),
-                    |decoded| Some(decoded.to_string()),
-                );
-
-                self.advance()?;
-
-                quasis.push(TemplateElement {
-                    raw: content,
-                    cooked,
-                    tail: true,
-                    span: Span::new(elem_start as u32, elem_end as u32),
-                });
-
-                Ok(Expression::TemplateLiteral(TemplateLiteral {
-                    quasis,
-                    expressions,
-                    span: Span::new(start as u32, elem_end as u32),
-                }))
-            }
-            TokenKind::TemplateHead => {
-                // Template with interpolation: `hello ${name}...`
-                let (elem_start, elem_end) = self.current_pos();
-                let raw = self.current_value();
-                let content = extract_template_head_content(raw).to_string();
-                let cooked = self.current_decoded().map_or_else(
-                    || Some(content.clone()),
-                    |decoded| Some(decoded.to_string()),
-                );
-
-                self.advance()?;
-
-                quasis.push(TemplateElement {
-                    raw: content,
-                    cooked,
-                    tail: false,
-                    span: Span::new(elem_start as u32, elem_end as u32),
-                });
-
-                self.grouping_depth += 1;
-
-                // Parse expressions and remaining template parts
-                loop {
-                    // Parse the interpolated expression
-                    let expr = self.parse_expression()?;
-                    expressions.push(expr);
-
-                    // Expect closing } of the interpolation
-                    let (brace_start, _) = self.current_pos();
-                    if !self.check(&TokenKind::BraceClose) {
-                        return Err(self.error_expected_found_at(
-                            "'}' at end of template interpolation",
-                            brace_start,
-                        ));
-                    }
-
-                    // Get the raw end position (without base_offset) for the lexer
-                    let raw_brace_end = self.current_raw_end();
-
-                    // Skip the } in the lexer without getting next token normally
-                    // (calling advance() would try to lex ` as a new token)
-                    // Instead, tell the lexer to skip past the } and read template content
-                    let next_token = self.lexer.continue_template_from_brace(raw_brace_end)?;
-                    self.update_current(next_token);
-
-                    let (elem_start, elem_end) = self.current_pos();
-                    let raw = self.current_value().to_string();
-
-                    match *self.current_kind() {
-                        TokenKind::TemplateMiddle => {
-                            // More interpolations to come: }content${
-                            let content = extract_template_head_content(&raw).to_string();
-                            let cooked = self.current_decoded().map_or_else(
-                                || Some(content.clone()),
-                                |decoded| Some(decoded.to_string()),
-                            );
-
-                            self.advance()?;
-
-                            quasis.push(TemplateElement {
-                                raw: content,
-                                cooked,
-                                tail: false,
-                                span: Span::new(elem_start as u32, elem_end as u32),
-                            });
-                        }
-                        TokenKind::TemplateTail => {
-                            // End of template: }content`
-                            let content = extract_template_tail_content(&raw).to_string();
-                            let cooked = self.current_decoded().map_or_else(
-                                || Some(content.clone()),
-                                |decoded| Some(decoded.to_string()),
-                            );
-
-                            self.advance()?;
-
-                            quasis.push(TemplateElement {
-                                raw: content,
-                                cooked,
-                                tail: true,
-                                span: Span::new(elem_start as u32, elem_end as u32),
-                            });
-
-                            break;
-                        }
-                        _ => {
-                            return Err(
-                                self.error_expected_found_at("template middle or tail", elem_start)
-                            );
-                        }
-                    }
-                }
-
-                self.grouping_depth -= 1;
-
-                let end = quasis.last().map_or(start as u32, |q| q.span.end);
-
-                Ok(Expression::TemplateLiteral(TemplateLiteral {
-                    quasis,
-                    expressions,
-                    span: Span::new(start as u32, end),
-                }))
-            }
-            _ => Err(self.error_expected_found_at("template literal", start)),
-        }
-    }
-
-    //
-    // Pattern Conversion (Cover Grammar)
-    //
-
-    /// Convert an expression to an assignable pattern (cover grammar)
-    ///
-    /// This implements the ECMAScript "cover grammar" for assignment targets.
-    /// When we parse `{a, b} = obj`, we first parse `{a, b}` as an ObjectExpression,
-    /// then convert it to an ObjectPattern when we see the `=`.
-    ///
-    /// Conversions:
-    /// - ObjectExpression → ObjectPattern
-    /// - ArrayExpression → ArrayPattern
-    /// - SpreadElement → RestElement
-    /// - BinaryExpression with = (shorthand default) → AssignmentPattern
-    /// - Identifier, MemberExpression → unchanged (valid assignment targets)
-    pub(super) fn to_assignable(&self, expr: Expression) -> Result<Expression, ParseError> {
-        match expr {
-            // Identifier is already a valid assignment target
-            Expression::Identifier(_) => Ok(expr),
-
-            // Member expression is a valid assignment target
-            Expression::MemberExpression(_) => Ok(expr),
-
-            // Convert ObjectExpression to ObjectPattern
-            Expression::ObjectExpression(obj) => {
-                let properties = obj
-                    .properties
-                    .into_iter()
-                    .map(|prop| self.object_property_to_pattern(prop))
-                    .collect::<Result<Vec<_>, _>>()?;
-
-                Ok(Expression::ObjectPattern(ObjectPattern {
-                    properties,
-                    type_annotation: None,
-                    span: obj.span,
-                }))
-            }
-
-            // Convert ArrayExpression to ArrayPattern
-            Expression::ArrayExpression(arr) => {
-                let elements = arr
-                    .elements
-                    .into_iter()
-                    .map(|elem| elem.map(|e| self.to_assignable(e)).transpose())
-                    .collect::<Result<Vec<_>, _>>()?;
-
-                Ok(Expression::ArrayPattern(ArrayPattern {
-                    elements,
-                    type_annotation: None,
-                    span: arr.span,
-                }))
-            }
-
-            // Convert SpreadElement to RestElement
-            Expression::SpreadElement(spread) => {
-                let argument = self.to_assignable(*spread.argument)?;
-                Ok(Expression::RestElement(RestElement {
-                    argument: Box::new(argument),
-                    type_annotation: None,
-                    span: spread.span,
-                }))
-            }
-
-            // AssignmentExpression in pattern context becomes AssignmentPattern
-            // This handles default values like `{a = 1}` which was parsed as shorthand
-            Expression::AssignmentExpression(assign) => {
-                let left = self.to_assignable(*assign.left)?;
-                Ok(Expression::AssignmentPattern(AssignmentPattern {
-                    left: Box::new(left),
-                    right: assign.right,
-                    span: assign.span,
-                }))
-            }
-
-            // Already a pattern (can happen with nested patterns)
-            Expression::ObjectPattern(_)
-            | Expression::ArrayPattern(_)
-            | Expression::AssignmentPattern(_)
-            | Expression::RestElement(_) => Ok(expr),
-
-            // Invalid assignment target
-            _ => Err(self.error_msg_at("Invalid assignment target", expr.span().start_usize())),
-        }
-    }
-
-    /// Convert an object property to a pattern property
-    fn object_property_to_pattern(
-        &self,
-        prop: ObjectProperty,
-    ) -> Result<ObjectPatternProperty, ParseError> {
-        match prop {
-            ObjectProperty::Property(p) => {
-                // Convert the value to a pattern
-                let value = self.to_assignable(p.value)?;
-
-                Ok(ObjectPatternProperty::Property(Property {
-                    key: p.key,
-                    value,
-                    method: p.method,
-                    shorthand: p.shorthand,
-                    computed: p.computed,
-                    kind: p.kind,
-                    span: p.span,
-                }))
-            }
-            ObjectProperty::SpreadElement(spread) => {
-                // Convert spread to rest element
-                let argument = self.to_assignable(*spread.argument)?;
-                Ok(ObjectPatternProperty::RestElement(RestElement {
-                    argument: Box::new(argument),
-                    type_annotation: None,
-                    span: spread.span,
-                }))
-            }
-        }
     }
 }
